+++ src/main/java/io/hops/metadata/hdfs/dal/EncodingStatusDataAccess.java
@@ -29,13 +29,13 @@ public interface EncodingStatusDataAccess<T> extends EntityDataAccess {
 
   void delete(T status) throws StorageException;
 
-  T findByInodeId(int inodeId) throws StorageException;
+  T findByInodeId(long inodeId) throws StorageException;
 
-  Collection<T> findByInodeIds(Collection<Integer> inodeIds) throws StorageException;
+  Collection<T> findByInodeIds(Collection<Long> inodeIds) throws StorageException;
   
-  T findByParityInodeId(int inodeId) throws StorageException;
+  T findByParityInodeId(long inodeId) throws StorageException;
 
-  Collection<T> findByParityInodeIds(List<Integer> inodeIds) throws StorageException;
+  Collection<T> findByParityInodeIds(List<Long> inodeIds) throws StorageException;
   
   Collection<T> findRequestedEncodings(int limit) throws StorageException;
