+++ src/main/java/io/hops/metadata/hdfs/entity/PendingBlockInfo.java
@@ -20,11 +20,11 @@ import java.util.List;
 public class PendingBlockInfo {
 
   private long blockId;
-  private int inodeId;
+  private long inodeId;
   private long timeStamp;
   private List<String> targets;
 
-  public PendingBlockInfo(long blockId, int inodeId, long timestamp,
+  public PendingBlockInfo(long blockId, long inodeId, long timestamp,
       List<String> targets) {
     this.blockId = blockId;
     this.inodeId = inodeId;
@@ -36,7 +36,7 @@ public class PendingBlockInfo {
     return blockId;
   }
 
-  public int getInodeId() {
+  public long getInodeId() {
     return inodeId;
   }
